Revert build always.

+++ bin/build-php-icu.sh
@@ -11,7 +11,7 @@ ICU_PHP_VERSION=$2
 
 ICU_DIR=$HOME/.build/icu-$ICU_VERSION
 ICU_PHP_DIR=$HOME/.build/php-$ICU_PHP_VERSION-icu-$ICU_VERSION
-#if [ ! -f $ICU_PHP_DIR/bin/php ]; then
+if [ ! -f $ICU_PHP_DIR/bin/php ]; then
 	wget -O icu-src.tgz http://download.icu-project.org/files/icu4c/$ICU_VERSION/icu4c-$(echo $ICU_VERSION | tr '.' '_')-src.tgz
 	mkdir icu-src && tar xzf icu-src.tgz -C icu-src --strip-components=1
 	pushd icu-src/source
@@ -24,4 +24,4 @@ ICU_PHP_DIR=$HOME/.build/php-$ICU_PHP_VERSION-icu-$ICU_VERSION
 		./configure --prefix=$ICU_PHP_DIR --enable-intl --with-icu-dir=$ICU_DIR --enable-mbstring --with-mysqli=/usr/bin/mysql_config --with-zlib --with-zlib-dir=/usr
 		make && make install
 	popd
-#fi
+fi
